requirements.txt
# Web Framework and Server
fastapi>=0.109.1
uvicorn[standard]==0.35.0

# Database
asyncpg==0.29.0
motor==3.3.2
pymongo==4.6.1

# Data Validation
pydantic>=2.11.7

# Authentication and Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Configuration
python-dotenv==1.1.0

# Security Enhancement Dependencies
cryptography>=41.0.0
bcrypt>=4.0.1